moved the `sql_database.py` from root folder to the `utilities/`
