+++ tests/unit_tests/agents/test_sql.py
@@ -1,6 +1,6 @@
 from langchain.agents import create_sql_agent
 from langchain.agents.agent_toolkits import SQLDatabaseToolkit
-from langchain.sql_database import SQLDatabase
+from langchain.utilities.sql_database import SQLDatabase
 from tests.unit_tests.llms.fake_llm import FakeLLM
